Add BottomNavigationBar and refactor
