+++ lib/calendar/calendar_page.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class CalendarPage extends StatefulWidget {
+  @override
+  _CalendarPageState createState() => _CalendarPageState();
+}
+
+class _CalendarPageState extends State<CalendarPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+      appBar: AppBar(
+        backgroundColor: Colors.orangeAccent,
+        title: Text('Calendar'),
+        centerTitle: true,
+      ),
+    );
+  }
+}
