+++ lib/dog/set_dog_data_page.dart
@@ -1,33 +1,79 @@
 import 'dart:convert';
+
+import 'package:datetime_picker_formfield/datetime_picker_formfield.dart';
+import 'package:doggo_frontend/globals.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:http/http.dart' as http;
+import 'package:intl/intl.dart';
 
-class SetUserDataState extends State<SetUserDataPage> {
+class SetDogDataPage extends StatefulWidget {
+  @override
+  _SetDogDataPageState createState() => _SetDogDataPageState();
+}
 
-  final firstNameController = TextEditingController();
-  final lastNameController = TextEditingController();
-  final hobbyController = TextEditingController();
+class _SetDogDataPageState extends State<SetDogDataPage> {
+  final nameController = TextEditingController();
+  final breedController = TextEditingController();
+  final colorController = TextEditingController();
+  final descriptionController = TextEditingController();
+  final vaccinationDateController = TextEditingController();
 
   @override
   void dispose() {
-    firstNameController.dispose();
-    lastNameController.dispose();
-    hobbyController.dispose();
+    nameController.dispose();
+    breedController.dispose();
+    colorController.dispose();
+    descriptionController.dispose();
+    vaccinationDateController.dispose();
     super.dispose();
+  }
+
+  Future showAlertDialogWithMessage(String message) {
+    return showDialog(
+        context: context,
+        builder: (context) {
+          return AlertDialog(content: Text(message));
+        });
+  }
+
+  Future setDogData() async {
+    final storage = FlutterSecureStorage();
+    final token = await storage.read(key: 'token');
+
+    var url = '$apiAddress/dogs';
+    var reqBody = jsonEncode({
+      'name': '${nameController.text}',
+      'breed': '${breedController.text}',
+      'color': '${colorController.text}',
+      'description': '${descriptionController.text}',
+      'lastVaccinationDate': '${vaccinationDateController.text}'
+    });
+    var headers = {
+      'Content-Type': 'application/json',
+      'Accept': '*/*',
+      'Authorization': 'Bearer $token'
+    };
+    final response = await http.post(url, body: reqBody, headers: headers);
+    if (response.statusCode == 200) {
+      Navigator.of(context).pushNamedAndRemoveUntil(
+          '/userroot', (Route<dynamic> route) => false);
+    } else {
+      showAlertDialogWithMessage('Could not set dog data!');
+    }
+    print(response.body);
   }
 
   @override
   Widget build(BuildContext context) {
-
     return Scaffold(
       appBar: AppBar(
-        title: Text('Set Your Details'),
+        title: Text('Set Dog\'s Details'),
       ),
       body: SingleChildScrollView(
         child: Container(
           alignment: Alignment.center,
-          margin: EdgeInsets.only(top: 30),
+          margin: EdgeInsets.only(top: 10),
           child: Column(
             children: <Widget>[
               Padding(
@@ -50,10 +96,10 @@ class SetUserDataState extends State<SetUserDataPage> {
                           Container(
                             padding: EdgeInsets.all(8),
                             child: TextField(
-                              controller: firstNameController,
+                              controller: nameController,
                               decoration: InputDecoration(
                                 border: InputBorder.none,
-                                hintText: "First name",
+                                hintText: "Name",
                                 hintStyle: TextStyle(color: Colors.grey),
                               ),
                             ),
@@ -62,46 +108,51 @@ class SetUserDataState extends State<SetUserDataPage> {
                           Container(
                             padding: EdgeInsets.all(8),
                             child: TextField(
-                              controller: lastNameController,
+                              controller: breedController,
                               decoration: InputDecoration(
                                 border: InputBorder.none,
-                                hintText: "Last name",
+                                hintText: "Breed",
                                 hintStyle: TextStyle(color: Colors.grey),
                               ),
                             ),
                           ),
                           Divider(color: Colors.grey),
                           Container(
-                              padding: EdgeInsets.all(8),
-                              child: DropdownButton(
-                                value: dropdownValue,
-                                icon: Icon(Icons.arrow_downward),
-                                iconSize: 16,
-                                hint: Text("Age"),
-                                style: TextStyle(color: Colors.orangeAccent),
-                                onChanged: (String newValue) {
-                                  setState(() {
-                                    dropdownValue = newValue;
-                                  });
-                                },
-                                items: dropdownMenuItems.map((String value) {
-                                  return DropdownMenuItem(
-                                    value: value,
-                                    child: Text(value),
-                                  );
-                                }).toList(),
-                              )),
+                            padding: EdgeInsets.all(8),
+                            child: TextField(
+                              controller: colorController,
+                              decoration: InputDecoration(
+                                border: InputBorder.none,
+                                hintText: "Color",
+                                hintStyle: TextStyle(color: Colors.grey),
+                              ),
+                            ),
+                          ),
                           Divider(color: Colors.grey),
                           Container(
                             padding: EdgeInsets.all(8),
                             child: TextField(
-                              controller: hobbyController,
+                              controller: descriptionController,
                               decoration: InputDecoration(
                                 border: InputBorder.none,
-                                hintText: "Hobby",
+                                hintText: "Description",
                                 hintStyle: TextStyle(color: Colors.grey),
                               ),
                             ),
+                          ),
+                          Divider(color: Colors.grey),
+                          Text('Click below to select vaccination date',
+                              style: TextStyle(color: Colors.grey)),
+                          DateTimeField(
+                            controller: vaccinationDateController,
+                            format: DateFormat("yyyy-MM-dd"),
+                            onShowPicker: (context, currentValue) {
+                              return showDatePicker(
+                                  context: context,
+                                  firstDate: DateTime(1900),
+                                  initialDate: currentValue ?? DateTime.now(),
+                                  lastDate: DateTime(2100));
+                            },
                           ),
                         ],
                       ),
@@ -113,7 +164,7 @@ class SetUserDataState extends State<SetUserDataPage> {
                       height: 50.0,
                       child: MaterialButton(
                         onPressed: () {
-                          addUserDetails();
+                          setDogData();
                         },
                         shape: RoundedRectangleBorder(
                             borderRadius: BorderRadius.circular(10)),
@@ -154,44 +205,4 @@ class SetUserDataState extends State<SetUserDataPage> {
       ),
     );
   }
-
-  Future addUserDetails() async {
-    final storage = FlutterSecureStorage();
-    final token = await storage.read(key: 'token');
-
-    var url = 'https://doggo-app-server.herokuapp.com/api/dogLover';
-    var body = jsonEncode({
-      'firstName': '${firstNameController.text}',
-      'lastName': '${lastNameController.text}',
-      'age': '$dropdownValue',
-      'hobby': '${hobbyController.text}'
-    });
-    var headers = {'Content-Type': 'application/json', 'Accept': '*/*', 'Authorization': 'Bearer $token'};
-
-    final response = await http.put(url, body: body, headers: headers);
-    if (response.statusCode == 200) {
-      Navigator.of(context).pushNamedAndRemoveUntil(
-          '/adddogdata', (Route<dynamic> route) => false
-      );
-    } else
-      showAlertDialogWithMessage('Could not set user data!');
-  }
-
-  Future showAlertDialogWithMessage(String message) {
-    return showDialog(
-        context: context,
-        builder: (context) {
-          return AlertDialog(content: Text(message));
-        });
-  }
-
-  String dropdownValue;
-
-  List<String> dropdownMenuItems =
-      List<String>.generate(99, (i) => (i + 1).toString());
-}
-
-class SetUserDataPage extends StatefulWidget {
-  @override
-  SetUserDataState createState() => SetUserDataState();
 }
