+++ lib/globals.dart
@@ -1,0 +1,1 @@
+const apiAddress = 'https://doggo-app-server.herokuapp.com/api';
