+++ lib/login_page.dart
@@ -1,6 +1,8 @@
+import 'dart:convert';
+
+import 'package:doggo_frontend/globals.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'dart:convert';
 import 'package:http/http.dart' as http;
 
 class LoginPageState extends State<LoginPage> {
@@ -135,7 +137,7 @@ class LoginPageState extends State<LoginPage> {
   Future loginUser() async {
     final storage = FlutterSecureStorage();
 
-    var postUrl = 'https://doggo-app-server.herokuapp.com/api/auth/signin';
+    var postUrl = '$apiAddress/auth/signin';
     var postRequestBody = jsonEncode({
       'email': '${emailController.text}',
       'password': '${passwordController.text}'
@@ -147,7 +149,7 @@ class LoginPageState extends State<LoginPage> {
     final token = jsonDecode(postResponse.body)['token'];
 
     if (postResponse.statusCode == 200) {
-      var getUrl = 'https://doggo-app-server.herokuapp.com/api/dogLover';
+      var getUrl = '$apiAddress/dogLover';
       var getHeaders = {'Authorization': 'Bearer $token'};
 
       await storage.write(key: 'token', value: token);
@@ -155,7 +157,7 @@ class LoginPageState extends State<LoginPage> {
 
       if (getResponse.statusCode == 200) {
         Navigator.of(context).pushNamedAndRemoveUntil(
-            '/userprofile', (Route<dynamic> route) => false);
+            '/userroot', (Route<dynamic> route) => false);
       } else if (getResponse.statusCode == 404) {
         Navigator.of(context).pushNamedAndRemoveUntil(
             '/adduserdata', (Route<dynamic> route) => false);
