+++ lib/main.dart
@@ -1,13 +1,13 @@
 import 'package:doggo_frontend/registration_page.dart';
+import 'package:doggo_frontend/user_main_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'Dog/dogs_list_page.dart';
-import 'Dog/edit_dog_data_page.dart';
-import 'Dog/set_dog_data_page.dart';
-import 'User/edit_user_profile_page.dart';
-import 'User/set_user_data_page.dart';
-import 'User/user_profile.dart';
+
+import 'dog/edit_dog_data_page.dart';
+import 'dog/set_dog_data_page.dart';
 import 'login_page.dart';
+import 'user/edit_user_profile_page.dart';
+import 'user/set_user_data_page.dart';
 
 void main() => runApp(MyApp());
 
@@ -17,19 +17,19 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'dogGO!',
       theme: ThemeData(
-          appBarTheme: AppBarTheme(color: Colors.orangeAccent),
-          brightness: Brightness.light),
+        appBarTheme: AppBarTheme(color: Colors.orangeAccent),
+        brightness: Brightness.light,
+      ),
       debugShowCheckedModeBanner: false,
       initialRoute: '/home',
       routes: {
         '/home': (context) => LoginPage(),
         '/register': (context) => RegistrationPage(),
-        '/userprofile': (context) => UserProfileView(),
-        '/dogsinfo': (context) => DogsListPage(),
+        '/userroot': (context) => UserMainPage(),
         '/adduserdata': (context) => SetUserDataPage(),
         '/adddogdata': (context) => SetDogDataPage(),
-        '/edituserdata': (context) => EditUserData(),
-        '/editdogdata': (context) => EditDogData(),
+        '/edituserdata': (context) => EditUserDataPage(),
+        '/editdogdata': (context) => EditDogDataPage(),
       },
     );
   }
