+++ lib/map/map_page.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class MapPage extends StatefulWidget {
+  @override
+  _MapPageState createState() => _MapPageState();
+}
+
+class _MapPageState extends State<MapPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+      appBar: AppBar(
+        backgroundColor: Colors.orangeAccent,
+        title: Text('Map'),
+        centerTitle: true,
+      ),
+    );
+  }
+}
