+++ lib/user/http/user_details_response.dart
